update (SMAPI 3.15.1, SDV 1.5.6)

+++ source/~daleao/ImmersiveValley/ImmersiveProfessions/Framework/Patches/Fishing/TownGetFishPatch.cs
@@ -64,7 +64,7 @@ internal sealed class TownGetFishPatch : DaLion.Common.Harmony.HarmonyPatch
                 .Advance()
                 .AddLabels(checkSeason)
                 .Insert(
-                    new CodeInstruction(OpCodes.Ldarg_S, 4) // arg 4 = Farmer who
+                    new CodeInstruction(OpCodes.Ldarg_S, (byte)4) // arg 4 = Farmer who
                 )
                 .InsertProfessionCheck(Profession.Angler.Value + 100, forLocalPlayer: false)
                 .Insert(
